==> 04-newdb.Rmd <==
```{r setup_newdb, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
eval_prevalences_savedb = FALSE
```

# Database for Calculator Shiny

In this chapter we are going to create database, which we can use in Shiny Applications as well in [results]()-section. In Shiny app user can add his input values for sleep apnea prevalences, condition prevalences and costs, and so get more specific results.


```{r}
library(duckdb)
library(arrow)
library(dplyr)
options(scipen = 999)

## Load duckdb
con = dbConnect(duckdb::duckdb(), dbdir="data/sleep22.duckdb", read_only=TRUE)

## Load from parquet files to duckdb mem
# if(FALSE){
#   con = dbConnect(duckdb::duckdb(), dbdir=":MEMORY:", read_only=FALSE)
#   fils <- list.files("data/parquet_database", full.names = T)
#   for (fi in fils) {
#     d <- arrow::read_parquet(file = fi)
#     nm <- gsub(pattern = "data/parquet_database/", replacement = "", x = fi)
#     nm <- gsub(pattern = ".parquet", replacement = "", x = nm)
#     duckdb::duckdb_register(conn = con, name = nm, df = d)
#   }
#   dbListTables(con)
# }
```

## Population datasets

Calculating 15-74 years old populations

```{r}
## 15-74yrs Both, female and male population in wide format
library(tidyr)
tbl(con, "pop") %>% 
  filter(age_group > 14 & age_group < 75 & year_id == 2019) %>% 
  group_by(location_name, sex_name) %>% 
  summarise(pop = sum(total_population)) %>% 
  collect() %>% 
  pivot_wider(names_from = sex_name, values_from = pop) %>% 
  rename(pop_both = both,
         pop_female = female,
         pop_male = male) %>% 
  mutate(
    pop_both = round(pop_both, 0),
    pop_female = round(pop_female, 0),
    pop_male = round(pop_male, 0),
    ages = "15-74"
  ) -> pop
head(pop)
```


Total populations

```{r}
## Total Both, female and male population in wide format
library(tidyr)
tbl(con, "pop") %>% 
  filter(year_id == 2019) %>% 
  group_by(location_name, sex_name) %>% 
  summarise(pop = sum(total_population)) %>% 
  collect() %>% 
  pivot_wider(names_from = sex_name, values_from = pop) %>% 
  rename(pop_both = both,
         pop_female = female,
         pop_male = male) %>% 
  mutate(
    pop_both = round(pop_both, 0),
    pop_female = round(pop_female, 0),
    pop_male = round(pop_male, 0),
    ages = "total"
    ) -> popu_info
head(popu_info)

## One dataset
pop <- pop %>% rbind(popu_info)
```



## Prevalences from ihme dataset

Calculating prevalences percentages by joining prevalences and population information

```{r}
## Get cause_id's which we can use from IHME dataset
tbl(con, "causes") %>% 
  filter(type == "prevalence") %>%
  select(cause_id) %>% 
  collect() -> causes
causes <- causes$cause_id[!is.na(causes$cause_id)]

## Get number of prevalences per location and causes from IHME (age: 15-74 and sex: both)
tbl(con, "ihme") %>% 
  filter(cause_id %in% causes & metric_name == "Number" & age_id > 7 & age_id < 20 & sex_id == 3) %>% 
  group_by(location_name, cause_id, cause_name) %>% 
  summarise(prevalence_number = sum(val)) %>% 
  arrange(location_name, cause_id) %>% 
  collect() -> prevalences_ihme
## Calculate Prevalence Rates: join tables prevalences_ihme & pop
prevalences_ihme %>% 
  right_join(pop %>% filter(ages == "15-74"), by = "location_name") %>% 
  mutate(prevalence = prevalence_number / pop_both
         )  %>% 
  select(location_name, cause_id, cause_name, prevalence) -> prevalences_ihme

head(prevalences_ihme)
```



## Mortalities from ihme

Calculating mortalities for two diseases and calculating percentages by joining mortality and population information

- 1023 	Other cardiovascular and circulatory diseases
- 294  	All causes

```{r}
## Get cause_id's which base values we can use to overwrite from IHME data
tbl(con, "causes") %>% 
  filter(type == "mortality") %>% 
  select(cause_id) %>% 
  collect() -> causes
causes <- causes$cause_id[!is.na(causes$cause_id)]

## Get mortality values from ihme
tbl(con, "ihme") %>% 
  filter(cause_id %in% causes & measure_name == "Deaths" & metric_name == "Number" & age_id > 7 & age_id < 20 & sex_id == 3) %>% 
  group_by(location_name, cause_id, cause_name) %>% 
  summarise(mortality_number = sum(val)) %>% 
  arrange(location_name, cause_id) %>% 
  collect() -> mortalities_ihme
## Calculate Rates: join tables mortalities_ihme & pop
mortalities_ihme %>% 
  right_join(pop %>% filter(ages == "15-74"), by = "location_name") %>% 
  mutate(mortality = mortality_number / pop_both) %>% 
  select(location_name, cause_id, cause_name, mortality) -> mortalities_ihme

head(mortalities_ihme)
```

```{r}
## Join mortalities and prevalance data. Add name to describe value and rename 'mortality'/'prevalence' to 'value'.
prevalences_ihme$type <- "prevalence"
mortalities_ihme$type <- "mortality"
ihme <- prevalences_ihme %>%
  rbind(mortalities_ihme) %>% 
  mutate(ihme = ifelse(is.na(prevalence), mortality, prevalence)) %>% 
  select(location_name, cause_id, cause_name, type, ihme) %>% 
  arrange(location_name, type, cause_id)
head(ihme)
```


```{r}
rm(list=c("mortalities_ihme", "prevalences_ihme"))
```



## Add also base values

Base data (from excel) and country specific prevalences & mortalities


```{r}
## Simple causes data / PAF and base values for each country
prevalences_base_simple <- NULL
for(cntry in unique(ihme$location_name)){
  tbl(con, "causes_simple") %>% 
    collect() %>% 
    mutate(location_name = cntry) %>%
    select("location_name", 
           "cause_id", 
           "condition", 
           "PAF", 
           "multiplier",
           "prevalence_base_italy", 
           "direct_healthcare_cost",
           "direct_nonhealthcare_cost", 
           "productivity_losses_cost") %>% 
    rbind(prevalences_base_simple) -> prevalences_base_simple
}
head(prevalences_base_simple) #1296

## Multirow causes data / PAF and base values for each country
prevalences_base <- NULL
for(cntry in unique(ihme$location_name)){
  # print(cntry)
  # tbl(con, "causes") %>% 
  tbl(con, "causes") %>% 
    collect() %>% 
    mutate(location_name = cntry) %>% 
    select("location_name", 
           "cause_id", 
           "condition", 
           "OSA_severity",
           "gender",
           "OR",
           "RR",
           "PAF", 
           "multiplier",
           "prevalence_base_italy", 
           "direct_healthcare_cost",
           "direct_nonhealthcare_cost", 
           "productivity_losses_cost") %>% 
    rbind(prevalences_base) -> prevalences_base
}
head(prevalences_base) #1296
# 1296 *2 2592
```

With base values, join new calculated prevalences/mortalities from IHME

```{r}
## Join base data and new calculated prevalences
## Calculate condition prevalences from ihme data using multiplier (=ratio between armeni and ihme condition)
prevalences <- prevalences_base %>%
  left_join(ihme, by = c("location_name", "cause_id")) %>% 
  select(location_name, condition, OSA_severity, gender, cause_id, cause_name, multiplier, ihme, prevalence_base_italy, OR, RR, PAF, direct_healthcare_cost, direct_nonhealthcare_cost, productivity_losses_cost) %>% 
  right_join(
    pop %>% 
      filter(ages == "15-74") %>% 
      select(location_name, pop_both, pop_female, pop_male),
    by = "location_name") %>%  ## Add the population to all rows
  mutate(#tamaluku = ihme,
         ihme = ifelse(is.na(ihme), NA, multiplier * ihme),
         multiplier = NULL)


rm(list = c("prevalences_base"))
```

And same to simple table 

```{r}
## Join base data and new calculated prevalences
## Calculate condition prevalences from ihme data using multiplier (=ratio between armeni and ihme condition)
prevalences_simple <- prevalences_base_simple %>%
  left_join(ihme, by = c("location_name", "cause_id")) %>% 
  select(location_name, condition, cause_id, cause_name, multiplier, ihme, prevalence_base_italy,  PAF, direct_healthcare_cost, direct_nonhealthcare_cost, productivity_losses_cost) %>% 
  right_join(
    pop %>% 
      filter(ages == "15-74") %>% 
      select(location_name, pop_both, pop_female, pop_male),
    by = "location_name") %>%  ## Add the population to all rows
  mutate(ihme = ifelse(is.na(ihme), NA, multiplier * ihme),
         multiplier = NULL)


head(prevalences_simple)

rm(list = c("prevalences_base_simple", "ihme"))
```




## Table: prevalences and mortalities

```{r}
DT::datatable(prevalences, rownames = FALSE)
```



## OSA values from article (Excel)

```{r}
### TODO left here, get prevelance data for slider
library(data.table)
## Read from source excel
osa <- readxl::read_xlsx("files/osa_tables.xlsx", sheet = "osa_armeni")
## Sleep apnea prevalences
osanew <- readxl::read_xlsx("files/osa_tables.xlsx", sheet = "osa_benjafield")
osanew$location_name
## How many countries are needed
length(unique(prevalences$location_name))
## Check which countries are not included in excel
length(osanew$location_name[osanew$location_name %in% unique(prevalences$location_name)])
## 3 countries missing, check names which are missing
unique(prevalences$location_name)[!unique(prevalences$location_name) %in% osanew$location_name]
## these are not included # [1] "Andorra"        "Monaco"         "San Marino"  

## Filter dataset and Calculate Mild, Moderate, Severe values separately
osanew <-osanew %>% 
  filter(location_name %in% unique(prevalences$location_name)) %>% 
  mutate(
    Mild = `Mild-Moderate-Severe` - `Moderate-Severe`,
    `Male_Moderate-Severe` = 1.3823 * `Moderate-Severe`, # Gender proportion is calculated from Italy data from Armeni article data
    `Female_Moderate-Severe` = 0.6508 * `Moderate-Severe`, # Gender proportion is calculated from Italy data from Armeni article data
    `Both_Moderate-Severe` = `Moderate-Severe`, # Gender proportion is calculated from Italy data from Armeni article data
    Both_Mild = Mild,
    Male_Mild =  0.9519 * Mild, # Gender proportion is calculated from Italy data from Armeni article data
    Female_Mild = 1.0440 * Mild # Gender proportion is calculated from Italy data from Armeni article data
  )

osanew <- osanew %>% 
  mutate(
    ## TODO check these with comments
    Female_Moderate = 0.5342 * `Female_Moderate-Severe`, # Moderate proportion percentage is calculated from Italy data from Armeni article data
    Female_Severe =  0.4658 * `Female_Moderate-Severe`,   # Severe proportion percentage is calculated from Italy data from Armeni article data
    Male_Moderate = 0.4004 * `Male_Moderate-Severe`, # Moderate proportion percentage is calculated from Italy data from Armeni article data
    Male_Severe =  0.5996 * `Male_Moderate-Severe` ,
    Both_Moderate = 0.445907 * `Both_Moderate-Severe`,
    Both_Severe = 0.554093 * `Both_Moderate-Severe`
  )

## To longer format
library(tidyr)
osanew <- osanew %>%
  tidyr::pivot_longer(c(`Male_Moderate-Severe`, `Male_Moderate`, `Male_Severe`, Male_Mild, `Female_Moderate-Severe`, `Female_Moderate`, `Female_Severe`,  Female_Mild, `Both_Moderate-Severe`, `Both_Moderate`, `Both_Severe`, Both_Mild), names_to = "OSA_severity") %>%
  rename(osa_rate=value) %>%
  mutate(
    ## Get gender and OSA_severity right
    gender = sub("_.*", "", OSA_severity),
    OSA_severity = sub('.+_(.+)', '\\1', OSA_severity) 
    ) %>% 
  select(location_name, OSA_severity, gender, osa_rate)
```

## Money index Excel

```{r}
## First try HICP index, what countries are included -----
## https://appsso.eurostat.ec.europa.eu/nui/show.do?dataset=prc_hicp_aind&lang=en
hicp <- readxl::read_xls("files/prc_hicp_aind.xls")
hicp$`GEO/TIME`[hicp$`GEO/TIME` == "Germany (until 1990 former territory of the FRG)"] <- "Germany" ## change name of the country

## Calculate correction index to all other countries
hicp %>% 
  filter(`GEO/TIME` %in% unique(prevalences$location_name)) %>% 
  rename(location_name = `GEO/TIME`) %>% 
  select(location_name, `2019`) %>% 
  mutate(index = `2019` / hicp$`2019`[hicp$`GEO/TIME` == "Italy"],
         ver = "EuroStat") %>% 
  select(location_name, index, ver) -> money_correction

## This is missing few countries so their index correction is 1 (equal to Italy)
## TODO find a way to correct these
# unique(prevalences_base$location_name[!prevalences_base$location_name %in% money_correction$location_name])
```


```{r}
dbDisconnect(con, shutdown=TRUE)
```


## Save data to new database

### duckdb

```{r, eval=eval_prevalences_savedb}
# library(duckdb)
# con = dbConnect(duckdb::duckdb(), dbdir="./data/sleep22_shiny.duckdb", read_only=FALSE)
# dbWriteTable(con, "prevalences", prevalences, overwrite=TRUE)
# dbWriteTable(con, "pop", pop, overwrite=TRUE)
# dbWriteTable(con, "osa", osa, overwrite=TRUE)
# dbWriteTable(con, "osanew", osanew, overwrite=TRUE)
# dbWriteTable(con, "money_correction", money_correction, overwrite=TRUE)
# dbDisconnect(con)
# 
# con = dbConnect(duckdb::duckdb(), dbdir="./data/sleep22_shiny.duckdb", read_only=FALSE)
# if(dir.exists("data/parquet_shiny/")) fs::dir_delete("data/parquet_shiny/")
# if(!dir.exists("data/parquet_shiny/")) dir.create("data/parquet_shiny/")
# dbSendQuery(con, "EXPORT DATABASE 'data/parquet_shiny/' (FORMAT PARQUET);") # copy komento? COPY taulu TO ... FORMAT 

# dbDisconnect(conn = con)
```

```{r, eval=FALSE}
## Copy newdb to sleep22calculator
# fils <- c(paste0(here::here(), "/data/sleep22_shiny.duckdb"), paste0(here::here(), "/data/sleep22_shiny.duckdb.wal"))
# fs::file_copy(fils, "/Users/japmiett/projects/sleep22calculator/", overwrite = T)
```

### Parquet



```{r}
## Testing another way to save database only to parquet files
con = dbConnect(duckdb::duckdb(), dbdir=":memory:", read_only=FALSE) ## TODO not working
## Register tbls
duckdb::duckdb_register(conn = con, name = "prevalences", df = prevalences)
duckdb::duckdb_register(conn = con, name = "prevalences_simple", df = prevalences_simple)
duckdb::duckdb_register(conn = con, name = "pop", df = pop)
duckdb::duckdb_register(conn = con, name = "osa", df = osa)
duckdb::duckdb_register(conn = con, name = "osanew", df = osanew)
duckdb::duckdb_register(conn = con, name = "money_correction", df = money_correction)
## Save to parquet
dbSendQuery(con, "COPY (SELECT * FROM prevalences) TO 'data/parquet_shiny/prevalences.parquet' (FORMAT 'parquet');")
dbSendQuery(con, "COPY (SELECT * FROM pop) TO 'data/parquet_shiny/pop.parquet' (FORMAT 'parquet');")
dbSendQuery(con, "COPY (SELECT * FROM osa) TO 'data/parquet_shiny/osa.parquet' (FORMAT 'parquet');")
dbSendQuery(con, "COPY (SELECT * FROM osanew) TO 'data/parquet_shiny/osanew.parquet' (FORMAT 'parquet');")
dbSendQuery(con, "COPY (SELECT * FROM money_correction) TO 'data/parquet_shiny/money_correction.parquet' (FORMAT 'parquet');")
dbSendQuery(con, "COPY (SELECT * FROM prevalences_simple) TO 'data/parquet_shiny/prevalences_simple.parquet' (FORMAT 'parquet');")
dbDisconnect(conn = con)
```

```{r}
## Copy to sleep22calculator
fils <- list.files("data/parquet_shiny/", full.names = T) 
fs::file_copy(fils, "/Users/japmiett/projects/sleep22calculator/data/", overwrite = T)
```








# Data for Visualizer Shiny

We need this function for calculating PAF from Odd Ratio. 

```{r}
## PAF ODDS RATIO Function -----
## Give only decimals in parameters
# OR = Odds Ratio
# PD = having a disease, prevalence
# PE = exposed, sleep apnea prevalence
# (PE_ =  unexposed)
paf_or <- function(OR, PD, PE){
  PD = PD * 100
  PE = PE * 100
  PE_ = 100 - PE
  VALUE1 = (PD * (1 - OR) + PE_ + OR * PE + sqrt( (PD * (1 - OR) + PE_ + OR * PE )^2 - 4 * PE_ * (1 - OR) *PD )) / (2 * PE_ * (1 - OR))
  VALUE2 = (PD * (1 - OR) + PE_ + OR * PE - sqrt( (PD * (1 - OR) + PE_ + OR * PE )^2 - 4 * PE_ * (1 - OR) *PD )) / (2 * PE_ * (1 - OR))
  VALUE <- ifelse(VALUE1 <= 100 & VALUE1 >= 0, VALUE1, VALUE2)
  PAF = 1 - ((100 * VALUE) / PD)
  return(PAF)
}


```

## Calculating main dataset

### Population

```{r}
population <- pop %>% 
  filter(ages=="15-74")
```


### OSA rates
```{r}
## Fixed multiplier values are calculated from original Armeni article table.
dosa <- osanew
slapnea_prevalence_female <- dosa$osa_rate[dosa$gender == "Female" & dosa$OSA_severity=="Moderate-Severe"]
slapnea_prevalence_male <- dosa$osa_rate[dosa$gender == "Male" & dosa$OSA_severity=="Moderate-Severe"]
# dosa$osa_rate[dosa$gender == "Female" & dosa$OSA_severity=="Moderate-Severe"] <- slapnea_prevalence_female / 100
dosa$osa_rate[dosa$gender == "Female" & dosa$OSA_severity=="Moderate"] <- 0.5342 * (slapnea_prevalence_female)
dosa$osa_rate[dosa$gender == "Female" & dosa$OSA_severity=="Severe"] <- 0.4658 * (slapnea_prevalence_female)
# dosa$osa_rate[dosa$gender == "Male" & dosa$OSA_severity=="Moderate-Severe"] <- slapnea_prevalence_male / 100
dosa$osa_rate[dosa$gender == "Male" & dosa$OSA_severity=="Moderate"] <-   0.4004  * (slapnea_prevalence_male)
dosa$osa_rate[dosa$gender == "Male" & dosa$OSA_severity=="Severe"] <-   0.5996 * (slapnea_prevalence_male)
dosa$osa_rate[dosa$gender == "Male" & dosa$OSA_severity=="Mild"] <- (slapnea_prevalence_male) / 1.44
dosa$osa_rate[dosa$gender == "Female" & dosa$OSA_severity=="Mild"] <- (slapnea_prevalence_female) / 0.6206897
dosa$osa_rate[dosa$gender == "Both" & dosa$OSA_severity=="Moderate-Severe"] <- (dosa$osa_rate[dosa$gender == "Female" & dosa$OSA_severity=="Moderate-Severe"] * population$pop_female + dosa$osa_rate[dosa$gender == "Male" & dosa$OSA_severity=="Moderate-Severe"] * population$pop_male) / population$pop_both
dosa$osa_rate[dosa$gender == "Both" & dosa$OSA_severity=="Moderate"] <- (dosa$osa_rate[dosa$gender == "Female" & dosa$OSA_severity=="Moderate"] * population$pop_female + dosa$osa_rate[dosa$gender == "Male" & dosa$OSA_severity=="Moderate"] * population$pop_male) / population$pop_both
dosa$osa_rate[dosa$gender == "Both" & dosa$OSA_severity=="Severe"] <- (dosa$osa_rate[dosa$gender == "Female" & dosa$OSA_severity=="Severe"] * population$pop_female + dosa$osa_rate[dosa$gender == "Male" & dosa$OSA_severity=="Severe"] * population$pop_male) / population$pop_both
dosa$osa_rate[dosa$gender == "Both" & dosa$OSA_severity=="Mild"] <- (dosa$osa_rate[dosa$gender == "Female" & dosa$OSA_severity=="Mild"] * population$pop_female + dosa$osa_rate[dosa$gender == "Male" & dosa$OSA_severity=="Mild"] * population$pop_male) / population$pop_both
## COMPARED TO FIXED ITALY VALUES AND ITS OK!
```

### Condition Prevalences and PAF

```{r}
## Prevalences and OSA rates added
d <- prevalences %>% 
  mutate(OSA_severity = ifelse(OSA_severity == "Overall", "Moderate-Severe", OSA_severity),
         prevalence = ifelse(is.na(ihme), prevalence_base_italy, ihme)) %>% 
  left_join(dosa,  by = c("location_name", "OSA_severity", "gender")) %>% 
  left_join(money_correction, by = "location_name") 

## Hard coding change, because Armeni used population of daytime sleepiness, we need to use fixed osa rate. This is  estimated
d$osa_rate[d$condition == "Car accidents"] <- 0.06885
d$osa_rate[d$condition == "Work accidents"] <- 0.06885

## Calculate PAFs
d <- d %>% 
  group_by(location_name) %>% 
      mutate(
        ## PAF calculation for Risk Ratio or Odds Ratio:
        PAFRR = ifelse(!is.na(RR), (osa_rate * (RR - 1) / (osa_rate * (RR - 1) + 1)), NA),
        PAFOR = ifelse(!is.na(OR), paf_or(OR, prevalence, osa_rate), NA),
        PAF = ifelse(is.na(PAFOR), ifelse(!is.na(PAFRR), PAFRR, 0), PAFOR),
        ## Prevalents per conditions 
        prevalent_cases = ifelse(gender=="Both", prevalence * pop_both, ifelse(gender=="Female", prevalence * pop_female, prevalence * pop_male)), 
        prevalent_cases_influenced_osa = PAF * prevalent_cases,
        ## Costs per conditions
        direct_cost = prevalent_cases_influenced_osa * direct_healthcare_cost,
        direct_non_healthcare_cost = prevalent_cases_influenced_osa * direct_nonhealthcare_cost,
        productivity_lost_cost = prevalent_cases_influenced_osa * productivity_losses_cost
      ) %>%
      mutate(direct_cost = ifelse(is.na(direct_cost), 0 , direct_cost),
             direct_non_healthcare_cost = ifelse(is.na(direct_non_healthcare_cost), 0 , direct_non_healthcare_cost),
             productivity_lost_cost = ifelse(is.na(productivity_lost_cost), 0 , productivity_lost_cost),
             total_costs = direct_cost + direct_non_healthcare_cost + productivity_lost_cost) 
```

```{r}
# cHeCK
# d %>% 
#   filter(location_name == "Finland") -> temp
# write.csv(temp, "temp_finland.csv")
```



```{r}
## Calculater sums per country
## TODO add money index
d <- d %>%
  group_by(location_name, pop_female, pop_male, pop_both, index) %>%
  summarise(direct_cost = sum(direct_cost, na.rm = T),
            direct_non_healthcare_cost = sum(direct_non_healthcare_cost, na.rm = T),
            productivity_lost_cost = sum(productivity_lost_cost, na.rm = T)) %>%
  ## OSA absolute values with separated moderate/severe calculation (OSA population to divide costs to).
  mutate(
    absolute_value_severe_moderate = ( (pop_female * dosa$osa_rate[dosa$OSA_severity == "Moderate-Severe" & dosa$gender == "Female" & dosa$location_name == location_name]) + (pop_male * dosa$osa_rate[dosa$OSA_severity == "Moderate-Severe" & dosa$gender == "Male"& dosa$location_name == location_name])), 
    absolute_value_mild = (pop_female * dosa$osa_rate[dosa$OSA_severity == "Mild" & dosa$gender == "Female"& dosa$location_name == location_name] + pop_male * dosa$osa_rate[dosa$OSA_severity == "Mild" & dosa$gender == "Male"& dosa$location_name == location_name] ),
    ## Costs per patients
    patient_direct_cost = direct_cost / absolute_value_severe_moderate,
    patient_nonhealthcare_cost = direct_non_healthcare_cost / absolute_value_severe_moderate,
    patient_productivity_cost = productivity_lost_cost / absolute_value_severe_moderate,
    patient_total_cost = patient_direct_cost + patient_nonhealthcare_cost + patient_productivity_cost,
    index = ifelse(is.na(index), 1, index)
  )
```



```{r}
## Write to calculation database
saveRDS(d, file = "data/slapnea22.RDS")
## Write to calculation database
# saveRDS(slapnea22_eurostat, file = "data/slapnea22_eurostat.RDS")
```


## Map data

```{r, eval = FALSE}
## Data for map
# Read this shape file with the rgdal library. 
library(rgdal)
world_spdf <- readOGR( 
  dsn= paste0(getwd(),"/data/map/") , 
  layer="TM_WORLD_BORDERS_SIMPL-0.3",
  verbose=FALSE
)

## Countries in our data
countries <- unique(d$location_name)
## Countries in world data
# sort(world_spdf$NAME)

## Rename few countries in world data and subset
world_spdf$NAME[world_spdf$NAME == "The former Yugoslav Republic of Macedonia"] <- "Macedonia"
world_spdf$NAME[world_spdf$NAME == "Republic of Moldova" ] <- "Moldova"
europe_spdf <- subset(world_spdf, world_spdf$NAME %in% countries)
# countries[!countries %in% europe_spdf$NAME]

## Write to calculation database
saveRDS(europe_spdf, file = "data/europe_spdf.RDS")
```

## Save and copy data

```{r, eval=FALSE}
## Copy to sleep22visualizer
fils <- c(paste0(here::here(), "/data/europe_spdf.RDS"), paste0(here::here(), "/data/slapnea22.RDS"), paste0(here::here(), "/data/slapnea22_eurostat.RDS"))
fs::file_copy(fils, "/Users/japmiett/projects/sleep22visualizer/data/", overwrite = T)
```





# Validation

Next we are going to validate the calculation to the article [Armeni et al. (2019) *Cost-of-illness study of Obstructive Sleep Apnea Syndrome (OSAS) in Italy*](https://cergas.unibocconi.eu/sites/default/files/files/Cost-of-illness-study-of-Obstructive-Sleep-Apnea-Syndrome-%2528OSAS%2529-in-Italy_Report%25281%2529.pdf). Because we use different data, results are going to be different,
.

## Population

Population in Armeni dataset is 

: Aged 15-74: Population |
|-----------|------------|  
| Female    | 22 708 659 |
| Male	    | 22 279 959 |
| Total	    | 44 988 618 |

In our dataset from IHME data, population is a bit larger:

```{r, eval = FALSE, echo=FALSE}
slapnea22 %>% 
  filter(location_name == "Italy") %>% 
  select(pop_female, pop_male) %>% 
  mutate(pop_total = pop_female + pop_male) %>% 
  tidyr::pivot_longer(c(pop_total, pop_female, pop_male), "Population") %>% 
  select(Population, value)
```

## Prevalences

**under construction**

Prevalences in the article are combined from different datasources (table 9, page 37). We have a bit different prevalences which are shown below:

```{r, include=FALSE, eval=FALSE}
# # Stroke tarkistus
# tbl(con, "ihme") %>% 
#   filter(cause_id == 494 & location_name == "Italy" & measure_name == "Prevalence" & metric_name == "Number" & age_id > 7 & age_id < 20 & sex_id == 3) %>% 
#   summarise(stroke = sum(val, na.rm = T))
# # 430529
# 
# ## Population
# tbl(con, "pop") %>% 
#   filter(location_name == "Italy" & sex_name == "both" & age_group > 14 & age_group < 75 & year_id == 2019) %>% 
#   summarise(population = sum(total_population))
# # 45304482
# 
# ## Stroke 15-74 vaestosta
# (430529/45304482) * 100
## eli  0.95 % vaestosta
## artikkelissa luku on 0.2 %, joten ilmeisemmin laskenta on oikein (5x kustannus)
```


```{r, eval=FALSE, echo=FALSE}
val <- readxl::read_xlsx("files/validation-italy.xlsx", sheet = "validation_summary")
DT::datatable(val, rownames = FALSE)
```
